+++ src/OperationalInsights/OperationalInsights/ChangeLog.md
@@ -20,7 +20,7 @@
 ## Upcoming Release
 * Added `-ForceDelete` option for `Remove-AzOperationalInsightsworkspace`
 * Added new cmdlet `Get-AzOperationalInsightsDeletedWorkspace`
-* Added new cmdlet `Recover-AzOperationalInsightsWorkspace`
+* Added new cmdlet `Restore-AzOperationalInsightsWorkspace`
 
 ## Version 2.2.0
 * Fixed bug PSWorkspace doesn't implement IOperationalInsightsWorkspace [#12135]